feat(home): refatoração da página inicial com novos componentes
