+++ boilerplate-nextjs-main/package.json
@@ -10,18 +10,23 @@
     "format": "prettier --write ."
   },
   "dependencies": {
+    "@tailwindcss/postcss": "^4.1.8",
     "lucide-react": "^0.511.0",
     "next": "^15.3.2",
+    "postcss": "^8.5.3",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "socket.io": "^4.8.1",
+    "socket.io-client": "^4.8.1"
   },
   "devDependencies": {
+    "@types/node": "^22.15.23",
     "@types/react": "19.1.6",
     "eslint": "^9.26.0",
     "eslint-config-next": "15.2.3",
     "lint-staged": "^15.2.2",
     "prettier": "^3.5.3",
-    "tailwindcss": "^4.1.5",
+    "tailwindcss": "^4.1.8",
     "typescript": "^5.8.3"
   },
   "lint-staged": {
