+++ boilerplate-nextjs-main/src/app/Home/CheckoutModal.tsx
@@ -1,0 +1,212 @@
+// app/(home)/Home/CheckoutModal.tsx
+"use client";
+import React, { useState, } from "react";
+import { Service } from "@/types/example";
+import { Shield, Users } from "lucide-react";
+
+export default function CheckoutModal({ isOpen, onClose, service }: { isOpen: boolean, onClose: () => void, service: Service | null }) {
+  const [formData, setFormData] = useState({
+    nome: "",
+    email: "",
+    telefone: "",
+    cpf: "",
+    cartao: "",
+    validade: "",
+    cvv: "",
+  });
+
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setFormData({ ...formData, [e.target.name]: e.target.value });
+  };
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    // Aqui seria a integração com o pagamento
+    alert("🎉 Contratação realizada com sucesso! Você receberá um email de confirmação em instantes.");
+    onClose();
+  };
+
+  if (!isOpen || !service) return null;
+
+
+  return (
+    <div className="fixed inset-0 bg-black/60 backdrop-blur-sm flex items-center justify-center z-50 p-4">
+      <div className="bg-white rounded-3xl shadow-2xl w-full max-w-2xl max-h-[90vh] overflow-y-auto">
+        <div className="bg-gradient-to-r from-blue-600 to-purple-600 text-white p-6 rounded-t-3xl">
+          <div className="flex justify-between items-center">
+            <div>
+              <h2 className="text-2xl font-bold">Finalizar Contratação</h2>
+              <p className="text-blue-100 text-sm">Últimos passos para sua proteção</p>
+            </div>
+            <button
+              onClick={onClose}
+              className="text-white hover:text-blue-200 text-2xl w-8 h-8 flex items-center justify-center rounded-full hover:bg-white/20 transition-all"
+            >
+              ✕
+            </button>
+          </div>
+        </div>
+        
+        <div className="p-6">
+          {/* Resumo do Produto Melhorado */}
+          <div className="bg-gradient-to-r from-blue-50 to-purple-50 p-6 rounded-2xl mb-6 border border-blue-200">
+            <div className="flex items-center gap-4 mb-4">
+              <div className="w-12 h-12 bg-gradient-to-r from-blue-600 to-purple-600 rounded-xl flex items-center justify-center">
+                <Shield className="w-6 h-6 text-white" />
+              </div>
+              <div>
+                <h3 className="font-bold text-slate-800 text-lg">{service.nome}</h3>
+                <p className="text-slate-600 text-sm">{service.descricao}</p>
+              </div>
+            </div>
+            <div className="flex justify-between items-center">
+              <div className="flex items-center gap-2">
+                <span className="text-3xl font-bold bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent">
+                  R$ {service.preco.toFixed(2).replace('.', ',')}
+                </span>
+                <span className="text-slate-500 text-sm">/mês</span>
+              </div>
+              <div className="bg-green-100 text-green-800 px-3 py-1 rounded-full text-sm font-medium">
+                🔒 Seguro e Protegido
+              </div>
+            </div>
+          </div>
+
+          {/* Formulário Melhorado */}
+          <form onSubmit={handleSubmit} className="space-y-6">
+            <div>
+              <h4 className="font-semibold text-slate-800 mb-4 flex items-center gap-2">
+                <Users className="w-5 h-5 text-blue-600" />
+                Dados Pessoais
+              </h4>
+              <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+                <div>
+                  <label className="block text-sm font-medium text-slate-700 mb-2">
+                    Nome Completo
+                  </label>
+                  <input
+                    type="text"
+                    name="nome"
+                    value={formData.nome}
+                    onChange={handleInputChange}
+                    required
+                    className="w-full px-4 py-3 border border-slate-300 rounded-xl focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-all"
+                    placeholder="Digite seu nome completo"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-slate-700 mb-2">
+                    Email
+                  </label>
+                  <input
+                    type="email"
+                    name="email"
+                    value={formData.email}
+                    onChange={handleInputChange}
+                    required
+                    className="w-full px-4 py-3 border border-slate-300 rounded-xl focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-all"
+                    placeholder="seu@email.com"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-slate-700 mb-2">
+                    Telefone
+                  </label>
+                  <input
+                    type="tel"
+                    name="telefone"
+                    value={formData.telefone}
+                    onChange={handleInputChange}
+                    required
+                    className="w-full px-4 py-3 border border-slate-300 rounded-xl focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-all"
+                    placeholder="(11) 99999-9999"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-slate-700 mb-2">
+                    CPF
+                  </label>
+                  <input
+                    type="text"
+                    name="cpf"
+                    value={formData.cpf}
+                    onChange={handleInputChange}
+                    required
+                    className="w-full px-4 py-3 border border-slate-300 rounded-xl focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-all"
+                    placeholder="000.000.000-00"
+                  />
+                </div>
+              </div>
+            </div>
+
+            <div className="border-t border-slate-200 pt-6">
+              <h4 className="font-semibold text-slate-800 mb-4 flex items-center gap-2">
+                💳 Dados do Cartão
+              </h4>
+              <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+                <div className="md:col-span-2">
+                  <label className="block text-sm font-medium text-slate-700 mb-2">
+                    Número do Cartão
+                  </label>
+                  <input
+                    type="text"
+                    name="cartao"
+                    value={formData.cartao}
+                    onChange={handleInputChange}
+                    placeholder="0000 0000 0000 0000"
+                    required
+                    className="w-full px-4 py-3 border border-slate-300 rounded-xl focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-all"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-slate-700 mb-2">
+                    Validade
+                  </label>
+                  <input
+                    type="text"
+                    name="validade"
+                    value={formData.validade}
+                    onChange={handleInputChange}
+                    placeholder="MM/AA"
+                    required
+                    className="w-full px-4 py-3 border border-slate-300 rounded-xl focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-all"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-slate-700 mb-2">
+                    CVV
+                  </label>
+                  <input
+                    type="text"
+                    name="cvv"
+                    value={formData.cvv}
+                    onChange={handleInputChange}
+                    placeholder="000"
+                    required
+                    className="w-full px-4 py-3 border border-slate-300 rounded-xl focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-all"
+                  />
+                </div>
+              </div>
+            </div>
+
+            <div className="flex gap-4 pt-6">
+              <button
+                type="button"
+                onClick={onClose}
+                className="flex-1 bg-slate-200 text-slate-800 py-4 rounded-xl font-semibold hover:bg-slate-300 transition-all"
+              >
+                Cancelar
+              </button>
+              <button
+                type="submit"
+                className="flex-1 bg-gradient-to-r from-blue-600 to-purple-600 text-white py-4 rounded-xl font-semibold hover:from-blue-700 hover:to-purple-700 transition-all duration-300 transform hover:scale-105 shadow-lg"
+              >
+                🚀 Finalizar Contratação
+              </button>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  );
+}
